Handle failed verify_auth requests by redirect without cookie.
This is good for logged-in users: They (may) get to see the image without further interaction.

This is fair when the head_sign cookie is missing for one reason or another: It is necessary to restart the login process.

This is also fair when the head_sign cookie does not match the claims, it is also necessary to restart.

Further improvement could include giving appropriate error descriptions